+++ packages/wallet/src/__stories__/index.tsx
@@ -6,7 +6,7 @@ import * as walletStates from '../redux/state';
 import * as channelStates from '../redux/channel/state';
 import * as fundingStates from '../redux/channel/funding/state';
 import '../index.scss';
-import * as scenarios from '../redux/reducers/__tests__/test-scenarios';
+import * as scenarios from '../redux/__tests__/test-scenarios';
 import { bigNumberify } from 'ethers/utils';
 import NetworkStatus from '../components/NetworkStatus';
 import { fundingConfirmed } from '../redux/channel/funding/state';
