+++ packages/wallet/src/redux/channel/__tests__/funding.test.ts
@@ -1,10 +1,10 @@
-import { fundingReducer } from '../../channel/funding/reducer';
+import { fundingReducer } from '../funding/reducer';
 
-import * as states from '../../channel/state';
-import * as fundingStates from '../../channel/funding/state';
+import * as states from '../state';
+import * as fundingStates from '../funding/state';
 import * as actions from '../../actions';
 
-import * as scenarios from './test-scenarios';
+import * as scenarios from '../../__tests__/test-scenarios';
 import {
   itTransitionsToChannelStateType,
   itIncreasesTurnNumBy,
@@ -13,15 +13,15 @@ import {
   expectThisCommitmentSent,
   itSendsThisTransaction,
   itSendsNoTransaction,
-} from './helpers';
+} from '../../__tests__/helpers';
 import * as TransactionGenerator from '../../../utils/transaction-generator';
 import * as outgoing from 'magmo-wallet-client/lib/wallet-events';
 import * as SigningUtil from '../../../utils/signing-utils';
 import * as fmgCore from 'fmg-core';
 import { bigNumberify } from 'ethers/utils';
 import { NextChannelState } from '../../sharedState';
-import { WAIT_FOR_FUNDING_REQUEST, WaitForFundingRequest } from '../../channel/funding/state';
-import { DIRECT_FUNDING } from '../../channel/funding/sharedState';
+import { WAIT_FOR_FUNDING_REQUEST, WaitForFundingRequest } from '../funding/state';
+import { DIRECT_FUNDING } from '../funding/sharedState';
 
 const {
   asAddress,
