+++ packages/wallet/src/redux/channel/__tests__/responding.test.ts
@@ -1,10 +1,10 @@
-import { respondingReducer } from '../../channel/respondingReducer';
+import { respondingReducer } from '../respondingReducer';
 
-import * as states from '../../channel/state';
+import * as states from '../state';
 import * as actions from '../../actions';
 
-import { itTransitionsToChannelStateType, itDoesntTransition } from './helpers';
-import * as scenarios from './test-scenarios';
+import { itTransitionsToChannelStateType, itDoesntTransition } from '../../__tests__/helpers';
+import * as scenarios from '../../__tests__/test-scenarios';
 import * as TransactionGenerator from '../../../utils/transaction-generator';
 import * as SigningUtil from '../../../utils/signing-utils';
 import * as FmgCore from 'fmg-core';
